requirements.txt
Django==1.5
psycopg2==2.4.6
django-extensions==1.1.0
ipython==0.13.1
simplejson==3.1.0
PIL==1.1.7
South==0.7.6
django-ajaxutils==0.2
